requirements.txt
asgiref==3.10.0
Django==5.2.7
django-phonenumber-field==8.3.0
phonenumbers==9.0.18
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2